Created Result class and implemented
Result class is made to handle API errors and responses

+++ Application/Posts/Delete.cs
@@ -1,4 +1,5 @@
-﻿using MediatR;
+﻿using Application.Core;
+using MediatR;
 using Persistence;
 using System;
 using System.Threading;
@@ -8,12 +9,12 @@ namespace Application.Posts
 {
     public class Delete
     {
-        public class Command : IRequest
+        public class Command : IRequest<Result<Unit>>
         {
             public Guid Id { get; set; }
         }
 
-        public class Handler : IRequestHandler<Command>
+        public class Handler : IRequestHandler<Command, Result<Unit>>
         {
             private readonly DataContext _context;
 
@@ -22,13 +23,20 @@ namespace Application.Posts
                 _context = context;
             }
 
-            public async Task<Unit> Handle(Command request, CancellationToken cancellationToken)
+            public async Task<Result<Unit>> Handle(Command request, CancellationToken cancellationToken)
             {
                 var postToDelete = await _context.Posts.FindAsync(request.Id);
-                _context.Posts.Remove(postToDelete);
-                await _context.SaveChangesAsync();
+                if (postToDelete == null) return null;
 
-                return Unit.Value;
+                _context.Posts.Remove(postToDelete);
+                var result = await _context.SaveChangesAsync() > 0;
+
+                if (!result)
+                {
+                    return Result<Unit>.Failure("Failed to delete post");
+                }
+
+                return Result<Unit>.Success(Unit.Value);
             }
         }
     }
